population of memory limit for the machine container
