+++ utils/machine/testdata/cpuinfo
